requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
pandas
sqlalchemy
jinja2
pydantic
python-dotenv
openai